how much knock would a knockback knock if a knock back could knock back
Ok i finally did the knocback but i only added it to stumpy but you can add it to any enemies by copying the code under stumpy and then adjusting the y axis and x axis knocback power as much as you want but i labeled it. Also i know its not perfect but with some animations and stuff it will look better. Plus it has a placeholder stutter effect so that makes it a little better. Hope you like it!